cmd/cover: show uncovered count in HTML dropdown
The HTML reports generated by cmd/cover show a percentage coverage for
each file in the file selector. This makes it hard to identify where
one can raise overall coverage in a project with a large number of
files. A file with 66% coverage but only a few lines is not nearly as
interesting as a file with the same percentage but thousands of lines.

This modifies the HTML report to also show a count of uncovered lines
next to the percentage, so that it's easier to spot files that might
need more testing attention.

Change-Id: I0939792bc0b1fc9b5d4eb4e12b5d5e95d997937b